Create separate folder for operators and add curl symbol
