+++ pararealml/core/mesh.py
@@ -1,112 +1,15 @@
-from abc import ABC, abstractmethod
 from copy import copy, deepcopy
 from typing import Tuple, Sequence
 
 import numpy as np
-from deepxde.geometry import Interval, Rectangle, Cuboid
-from deepxde.geometry.geometry import Geometry
 
 SpatialDomainInterval = Tuple[float, float]
 
 
-class Mesh(ABC):
+class Mesh:
     """
-    A mesh representing a discretised domain of arbitrary dimensionality.
-    """
-
-    @property
-    @abstractmethod
-    def x_intervals(self) -> Tuple[SpatialDomainInterval, ...]:
-        """
-        Returns the bounds of each axis of the domain
-        """
-
-    @property
-    @abstractmethod
-    def d_x(self) -> Tuple[float, ...]:
-        """
-        Returns the step sizes along the dimensions of the domain.
-        """
-
-    @property
-    @abstractmethod
-    def vertices_shape(self) -> Tuple[int, ...]:
-        """
-        Returns the shape of the array of the vertices of the discretised
-        domain.
-        """
-
-    @property
-    @abstractmethod
-    def cells_shape(self) -> Tuple[int, ...]:
-        """
-        Returns the shape of the array of the cell centers of the discretised
-        domain.
-        """
-
-    @property
-    @abstractmethod
-    def vertex_coordinates(self) -> Tuple[np.ndarray, ...]:
-        """
-        Returns a tuple of the coordinates of the vertices of the mesh along
-        each axis.
-        """
-
-    @property
-    @abstractmethod
-    def cell_center_coordinates(self) -> Tuple[np.ndarray, ...]:
-        """
-        Returns a tuple of the coordinates of the cell centers of the mesh
-        along each axis.
-        """
-
-    @property
-    @abstractmethod
-    def deepxde_geometry(self) -> Geometry:
-        """
-        Returns the DeepXDE equivalent of the spatial domain represented by the
-        mesh.
-        """
-
-    @abstractmethod
-    def x(self, index: Tuple[int, ...], vertex: bool) -> Tuple[float, ...]:
-        """
-        Returns the coordinates of the point in the domain corresponding to the
-        vertex or cell center of the mesh specified by the provided index.
-
-        :param index: the index of a vertex or cell center of the mesh
-        :param vertex: whether the point is a vertex or a cell center
-        :return: the coordinates of the corresponding point of the domain
-        """
-
-    def shape(self, vertex_oriented: bool):
-        """
-        Returns the shape of the array of the discretised domain.
-
-        :param vertex_oriented: whether the shape of the vertices or the cells
-            of the mesh is to be returned
-        :return: the shape of the vertices or the cells
-        """
-        return self.vertices_shape if vertex_oriented else self.cells_shape
-
-    def coordinates(self, vertex_oriented: bool) -> Tuple[np.ndarray, ...]:
-        """
-        Returns a tuple of the coordinates of the vertices or cell centers
-        of the mesh along each axis.
-
-        :param vertex_oriented: whether the coordinates of the vertices or the
-            cells of the mesh is to be returned
-        :return: a tuple of arrays each representing the coordinates along the
-            corresponding axis
-        """
-        return self.vertex_coordinates if vertex_oriented \
-            else self.cell_center_coordinates
-
-
-class UniformGrid(Mesh):
-    """
-    A rectangular grid of arbitrary dimensionality and shape with a uniform
-    spacing of grid points along each axis.
+    A hypercube of arbitrary dimensionality and shape with a uniform spacing of
+    grid points along each axis.
     """
 
     def __init__(
@@ -140,43 +43,90 @@ class UniformGrid(Mesh):
         self._x_cell_center_offset = np.array(
             [coordinates[0] for coordinates in self._cell_center_coordinates])
 
-        self._deepxde_geometry = self._create_deepxde_geometry()
-
     @property
     def x_intervals(self) -> Tuple[SpatialDomainInterval, ...]:
+        """
+        Returns the bounds of each axis of the domain
+        """
         return deepcopy(self._x_intervals)
 
     @property
     def d_x(self) -> Tuple[float, ...]:
+        """
+        Returns the step sizes along the dimensions of the domain.
+        """
         return tuple(self._d_x)
 
     @property
     def vertices_shape(self) -> Tuple[int, ...]:
+        """
+        Returns the shape of the array of the vertices of the discretised
+        domain.
+        """
         return copy(self._vertices_shape)
 
     @property
     def cells_shape(self) -> Tuple[int, ...]:
+        """
+        Returns the shape of the array of the cell centers of the discretised
+        domain.
+        """
         return copy(self._cells_shape)
 
     @property
     def vertex_coordinates(self) -> Tuple[np.ndarray, ...]:
+        """
+        Returns a tuple of the coordinates of the vertices of the mesh along
+        each axis.
+        """
         return deepcopy(self._vertex_coordinates)
 
     @property
     def cell_center_coordinates(self) -> Tuple[np.ndarray, ...]:
+        """
+        Returns a tuple of the coordinates of the cell centers of the mesh
+        along each axis.
+        """
         return deepcopy(self._cell_center_coordinates)
 
-    @property
-    def deepxde_geometry(self) -> Geometry:
-        return self._deepxde_geometry
-
     def x(self, index: Tuple[int, ...], vertex: bool) -> Tuple[float, ...]:
+        """
+        Returns the coordinates of the point in the domain corresponding to the
+        vertex or cell center of the mesh specified by the provided index.
+
+        :param index: the index of a vertex or cell center of the mesh
+        :param vertex: whether the point is a vertex or a cell center
+        :return: the coordinates of the corresponding point of the domain
+        """
         if len(index) != len(self._x_intervals):
             raise ValueError
 
         offset = self._x_vertex_offset if vertex \
             else self._x_cell_center_offset
         return tuple(offset + self._d_x * index)
+
+    def shape(self, vertex_oriented: bool):
+        """
+        Returns the shape of the array of the discretised domain.
+
+        :param vertex_oriented: whether the shape of the vertices or the cells
+            of the mesh is to be returned
+        :return: the shape of the vertices or the cells
+        """
+        return self.vertices_shape if vertex_oriented else self.cells_shape
+
+    def coordinates(self, vertex_oriented: bool) -> Tuple[np.ndarray, ...]:
+        """
+        Returns a tuple of the coordinates of the vertices or cell centers
+        of the mesh along each axis.
+
+        :param vertex_oriented: whether the coordinates of the vertices or the
+            cells of the mesh is to be returned
+        :return: a tuple of arrays each representing the coordinates along the
+            corresponding axis
+        """
+        return self.vertex_coordinates if vertex_oriented \
+            else self.cell_center_coordinates
 
     def _calculate_shape(
             self,
@@ -229,21 +179,3 @@ class UniformGrid(Mesh):
                 np.linspace(x_low, x_high, mesh_shape[i]))
 
         return tuple(coordinates)
-
-    def _create_deepxde_geometry(self) -> Geometry:
-        """
-        Creates and returns the DeepXDE equivalent of the spatial domain
-        represented by the mesh.
-        """
-        x_dimension = len(self._x_intervals)
-        if x_dimension == 1:
-            x_interval = self._x_intervals[0]
-            geometry = Interval(*x_interval)
-        elif x_dimension == 2:
-            geometry = Rectangle(*zip(*self._x_intervals))
-        elif x_dimension == 3:
-            geometry = Cuboid(*zip(*self._x_intervals))
-        else:
-            raise NotImplementedError
-
-        return geometry
